updated user login info

+++ Controller/users.js
@@ -1,40 +1,48 @@
+const dynamoClient = require("../helpers/awsconnetion");
 
-const bcrypt = require("bcrypt");
-const uuid = require("uuid");
-
-const dynamoClient = require('../helpers/awsconnetion')
-
-
-
-const TABLE_NAME = "RegisterTable";
+const TABLE_NAME = "UsersTable";
+const TABLE_NAME1 = "AdminTable";
 
 //REGISTER
 
-  // CREATE OR UPDATE USER
-  const createOrUpdateUser = async (data = {}) => {
-    const params = {
-      TableName: TABLE_NAME,
-      Item: data
-    }
-    try {
-      await dynamoClient.put(params).promise()
-      return { success: true }
-    } catch(error) {
-      return { success: false }
-    }
+// CREATE OR UPDATE USER
+const createOrUpdateUser = async (data = {}) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Item: data,
+  };
+  try {
+    await dynamoClient.put(params).promise();
+    return { success: true };
+  } catch (error) {
+    return { success: false };
   }
-
+};
 
 // FETCH ALL USERS
-const getAllUsers = async () => {
+const getAllUsers = async (email) => {
+  //query to check the db for email value
+  const scanParams = {
+    TableName: TABLE_NAME1,
+    FilterExpression: "email = :email",
+    ExpressionAttributeValues: {
+      ":email": { S: email },
+    },
+  };
+
   const params = {
     TableName: TABLE_NAME,
   };
   try {
+    //check if user is an admin to access this route
+    const data = await dynamoClient.scan(scanParams).promise();
+    if (data.Items.length === 0)
+      return { success: false, error: "Permission denied!!" };
+
     const { Items = [] } = await dynamoClient.scan(params).promise();
-    
+
     // Remove "S" and "N" attributes from the data
-    const transformedData = Items.map(item => {
+    const transformedData = Items.map((item) => {
       const transformedItem = {};
       for (const [key, value] of Object.entries(item)) {
         const attributeValue = Object.values(value)[0];
@@ -42,13 +50,12 @@ const getAllUsers = async () => {
       }
       return transformedItem;
     });
-    
+
     return { success: true, data: transformedData };
   } catch (error) {
     return { success: false, data: null };
   }
 };
-
 
 // READ SINGLE USER ON KEY(id)
 const getUser = async (value, key = "id") => {
@@ -60,30 +67,43 @@ const getUser = async (value, key = "id") => {
   };
   try {
     const { Item = {} } = await dynamoClient.get(params).promise();
-    
+
     // Remove "S" and "N" attributes from the data
     const transformedItem = {};
     for (const [key, value] of Object.entries(Item)) {
       const attributeValue = Object.values(value)[0];
       transformedItem[key] = attributeValue;
     }
-    
+
     return { success: true, data: transformedItem };
   } catch (error) {
     return { success: false, data: null };
   }
 };
 
-
 // Delete Existing User
-const deleteUser = async (value, key = "id") => {
+const deleteUser = async (user, key = "id") => {
+  //query to check the db for email value
+  const scanParams = {
+    TableName: TABLE_NAME1,
+    FilterExpression: "email = :email",
+    ExpressionAttributeValues: {
+      ":email": { S: user.email },
+    },
+  };
+
   const params = {
     TableName: TABLE_NAME,
     Key: {
-      [key]: value,
+      [key]: value.id,
     },
   };
   try {
+    //check if user is an admin to access this route
+    const data = await dynamoClient.scan(scanParams).promise();
+    if (data.Items.length === 0)
+      return { success: false, error: "Permission denied!!" };
+
     await dynamoClient.delete(params).promise();
     return { success: true };
   } catch (error) {
